Add Displayable Interface, write Presenter Tests
While trying to write tests for my Presenter, I realized
that Display needed an interface so I could make a
TestDisplay. As a result, Display no longer needs to be
initialized with an output. I also called Board#setUpGrid
in the constructor, so I don't need to fill the grid in
separately.